requirements.txt
Django==1.4.1
Markdown==2.2.0
PyYAML==3.10
